requirements.txt
# Web Framework
streamlit==0.79.0

# OpenAI
openai==0.10.2

# Azure
azure-storage-blob==12.8.0

# Google Auth
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-auth-httplib2==0.0.4

# Audio Processing
pydub==0.24.1

# Video Processing
youtube_dl==2021.3.14

# Other
requests==2.25.1
python-dotenv==0.15.0
